+++ CHANGELOG.rst
@@ -56,6 +56,11 @@ Deprecated
 - ``ArgumentParser.merge_config`` is deprecated and will be removed in v5.0.0.
   There is no replacement since it is considered internal (`#925
   <https://github.com/omni-us/jsonargparse/pull/925>`__).
+-  ``Optional`` type parameters without a default and required parameters
+   without a type annotation when using ``fail_untyped=False`` are currently set
+   to optional with default ``None``. In v5.0.0 they will be required. Warnings
+   visible with ``JSONARGPARSE_DEPRECATION_WARNINGS=all`` (`#929
+  <https://github.com/omni-us/jsonargparse/pull/929>`__).
 
 
 v4.49.0 (2026-05-15)